Updated how SpiceException is raised.
When a SpiceException was being raised, a new Exception object was being
created every time.  This did not allow the Python side to correlate that
exception with the SpiceException object in the module since they were distinct
objects.

This patch creates a global reference for SpiceException and it's the one being
thrown when an error occurs.